requirements.txt
# ============================================================
#  RideFlow Analytics Platform — Python Dependencies
# ============================================================

# Core PySpark / Delta
pyspark==3.5.0
delta-spark==3.1.0

# Kafka
confluent-kafka==2.3.0
kafka-python==2.0.2

# AWS
boto3==1.34.0
botocore==1.34.0

# Data utilities
faker==22.0.0
pandas==2.1.4
pyarrow==14.0.2

# Airflow (install separately or via pip extras)
apache-airflow==2.8.0
apache-airflow-providers-databricks==6.3.0
apache-airflow-providers-apache-kafka==1.3.0

# Monitoring / Logging
python-json-logger==2.0.7
slack-sdk==3.26.2

# Testing
pytest==7.4.4
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
